Feature/refactor campaign

+++ src/main/java/com/maliag/grimoireLink/features/campaign/service/CampaignServiceImpl.java
@@ -1,11 +1,15 @@
-package com.maliag.grimoireLink.features.campaign;
+package com.maliag.grimoireLink.features.campaign.service;
 
+import com.maliag.grimoireLink.common.exceptions.UnauthorizedException;
+import com.maliag.grimoireLink.features.campaign.model.CampaignEntity;
+import com.maliag.grimoireLink.features.campaign.mapper.CampaignMapper;
+import com.maliag.grimoireLink.features.campaign.repository.CampaignRepository;
 import com.maliag.grimoireLink.features.campaign.dto.CampaignRequest;
 import com.maliag.grimoireLink.features.campaign.dto.CampaignResponse;
 import com.maliag.grimoireLink.features.campaign.dto.UpdateCampaignRequest;
 
+import com.maliag.grimoireLink.features.campaign.exceptions.AlreadyMemberException;
 import com.maliag.grimoireLink.features.campaign.exceptions.CampaignNotFoundException;
-import com.maliag.grimoireLink.features.encounter.dto.EncounterResponse;
 import com.maliag.grimoireLink.features.users.models.UserEntity;
 import com.maliag.grimoireLink.features.users.repositories.UserRepository;
 import com.maliag.grimoireLink.features.users.services.UserService;
@@ -16,7 +20,6 @@ import com.maliag.grimoireLink.features.usersXCampaign.UsersXCampaignRepository;
 import com.maliag.grimoireLink.features.usersXCampaign.dto.CampaignMemberResponse;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,14 +30,12 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class CampaignServiceImpl implements CampaignService{
+public class CampaignServiceImpl implements CampaignService {
 
     private final CampaignRepository repository;
     private final CampaignMapper mapper;
     private final UsersXCampaignRepository uxcRepository;
     private final UsersXCampaignMapper uxcMapper;
-
-    private final UserRepository userRepository;
     private final UserService userService;
 
     @Transactional(readOnly = true)
@@ -50,8 +51,9 @@ public class CampaignServiceImpl implements CampaignService{
 
     @Override
     @Transactional(readOnly = true)
-    public List<CampaignResponse> getAllCampaignsByUserId(UUID userId){
-        return uxcRepository.findCampaignsByUserPublicId(userId).stream()
+    public List<CampaignResponse> getAllCampaignsByUser(){
+        UserEntity user = userService.getLoggedUserEntity();
+        return uxcRepository.findCampaignsByUserPublicId(user.getPublicId()).stream()
                 .map(mapper::toResponse)
                 .collect(Collectors.toList());
     }
@@ -93,6 +95,12 @@ public class CampaignServiceImpl implements CampaignService{
     @Transactional
     public CampaignResponse updateCampaign(UUID publicId, UpdateCampaignRequest request){
         CampaignEntity campaign = findByPublicId(publicId);
+        UserEntity user = userService.getLoggedUserEntity();
+
+        if(!getDm(publicId).getUser().getPublicId().equals(user.getPublicId())){
+            throw new UnauthorizedException("Only the DM can update the campaign");
+        }
+
         if(request.getName()!=null){
             campaign.setName(request.getName());
         }
@@ -109,6 +117,12 @@ public class CampaignServiceImpl implements CampaignService{
     @Transactional
     public void deleteCampaign(UUID publicId){
         CampaignEntity campaign = findByPublicId(publicId);
+        UserEntity user = userService.getLoggedUserEntity();
+
+        if(!getDm(publicId).getUser().getPublicId().equals(user.getPublicId())){
+            throw new UnauthorizedException("Only the DM can delete the campaign");
+        }
+
         ///usar baja logica?
         repository.delete(campaign);
     }
@@ -118,11 +132,11 @@ public class CampaignServiceImpl implements CampaignService{
 
         UserEntity user = userService.getLoggedUserEntity();
         CampaignEntity campaign = repository.findByInviteCode(inviteCode)
-                .orElseThrow(() -> new EntityNotFoundException("Código de invitación inválido"));
+                .orElseThrow(() -> new CampaignNotFoundException("Código de invitación inválido"));
 
         boolean alreadyMember = uxcRepository.existsByUserAndCampaign(user, campaign);
         if (alreadyMember) {
-            throw new IllegalStateException("Ya sos miembro de esta campaña");
+            throw new AlreadyMemberException("Ya sos miembro de esta campaña");
         }
 
         UsersXCampaignEntity membership = UsersXCampaignEntity.builder()
@@ -136,6 +150,11 @@ public class CampaignServiceImpl implements CampaignService{
         return mapper.toResponse(campaign);
     }
 
-
+    @Transactional(readOnly = true)
+    public UsersXCampaignEntity getDm(UUID campaignPublicId){
+        return uxcRepository
+                .findByCampaign_PublicIdAndRole(campaignPublicId, Role.DUNGEON_MASTER)
+                .orElseThrow(() -> new EntityNotFoundException("DM not found"));
+    }
 
 }
